requirements-file.txt
discord.py==2.3.2
PyNaCl==1.5.0
PyAudio==0.2.13
Flask==2.3.2
SQLAlchemy==2.0.19
aiohttp==3.8.5